# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://task.firscoders.shop/api/",
            "description": "Global API server"
        },
        {
            "url": "http://test-task.loc/api/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/v1/product-warehouses": {
            "get": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouses list",
                "description": "Product warehouses list.",
                "operationId": "productWarehousesList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "15"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouse add",
                "description": "Warehouse add.",
                "operationId": "productWarehousesCreate",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "description": "Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "43"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/product-warehouses/{product_warehouse_id}": {
            "get": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouse show",
                "description": "Product Warehouse show.",
                "operationId": "productWarehousesshow",
                "parameters": [
                    {
                        "name": "product_warehouse_id",
                        "in": "path",
                        "description": "Product Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Product Warehouse not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouse update",
                "description": "Product Warehouse update.",
                "operationId": "productWarehousesUpdate",
                "parameters": [
                    {
                        "name": "product_warehouse_id",
                        "in": "path",
                        "description": "Product Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Quantity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "43"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouse delete",
                "description": "Warehouse delete.",
                "operationId": "productWarehousesDelete",
                "parameters": [
                    {
                        "name": "product_warehouse_id",
                        "in": "path",
                        "description": "Product Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "3"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Product Warehouse not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/product-warehouses/{product_warehouse_id}/edit": {
            "get": {
                "tags": [
                    "Product Warehouses Routes"
                ],
                "summary": "Product Warehouse edit info",
                "description": "Product Warehouse edit info.",
                "operationId": "productWarehousesEdit",
                "parameters": [
                    {
                        "name": "product_warehouse_id",
                        "in": "path",
                        "description": "Product Warehouse id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}